Staking math upgrades

+++ contracts/staking/contracts/src/interfaces/IStructs.sol
@@ -41,10 +41,10 @@ interface IStructs {
 
     /// @dev State for Staking Pools (see MixinStakingPool).
     /// @param operatorAddress Address of pool operator.
-    /// @param operatorShare Portion of pool rewards owned by operator.
+    /// @param operatorShare Portion of pool rewards owned by operator, in ppm.
     struct Pool {
         address payable operatorAddress;
-        uint8 operatorShare;
+        uint32 operatorShare;
     }
 
     /// @dev State for a pool that actively traded during the current epoch.
